Prevent excessive rebuilds for modules
An order only dependency should be enough for depending on the module
dyndep file. CMake does the same. However CMake also has an additional
dependency on the per file modmap file. We don't have an equivalent
right now and it is not quite clear why that would be needed.

See https://gitlab.kitware.com/cmake/cmake/-/issues/25511 and
https://gitlab.kitware.com/cmake/cmake/-/merge_requests/9248 for more
information.